Age Pyramid

Age Pyramid

Usage

agepyramid(
  data,
  age,
  gender,
  female,
  bin_width = 5,
  plot_title = "Age Pyramid",
  color1 = "#1F77B4",
  color2 = "#FF7F0E",
  color_palette = "standard",
  age_groups = "custom",
  plot_engine = "ggcharts"
)

Arguments

data

The data as a data frame.

age

a string naming the variable from data that contains the continuous values used for the report

gender

a string naming the variable from data that contains the categorical values used for the report

female

a string naming the level from gender that contains the level female

bin_width

The width of the age bins in years. Adjust this to change the granularity of the age groups.

plot_title

The title displayed on the age pyramid plot.

color1

First color for the pyramid chart bars.

color2

Second color for the pyramid chart bars.

color_palette

Choose color palette type. Standard uses traditional gender colors, Accessible is colorblind-friendly, Custom allows user-defined colors.

age_groups

Preset age groupings for common clinical contexts. Custom uses bin_width setting.

plot_engine

Choose the plotting engine. ggcharts provides better pyramid styling, ggplot2 offers more customization.

Value

A results object containing:

results$pyramidTable					a table

results$plot					an image

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$pyramidTable$asDF

as.data.frame(results$pyramidTable)