Agreement

This function uses a dataframe to calculate decision tests

Value

list

Super classes

jmvcore::Analysis -> ClinicoPath::agreementBase -> agreementClass

Methods

Public methods

agreementClass$clone()

Inherited methods

jmvcore::Analysis$.createImage()

jmvcore::Analysis$.createImages()

jmvcore::Analysis$.createPlotObject()

jmvcore::Analysis$.load()

jmvcore::Analysis$.render()

jmvcore::Analysis$.save()

jmvcore::Analysis$.savePart()

jmvcore::Analysis$.setCheckpoint()

jmvcore::Analysis$.setParent()

jmvcore::Analysis$.setReadDatasetHeaderSource()

jmvcore::Analysis$.setReadDatasetSource()

jmvcore::Analysis$.setResourcesPathSource()

jmvcore::Analysis$.setStatePathSource()

jmvcore::Analysis$addAddon()

jmvcore::Analysis$asProtoBuf()

jmvcore::Analysis$asSource()

jmvcore::Analysis$check()

jmvcore::Analysis$init()

jmvcore::Analysis$optionsChangedHandler()

jmvcore::Analysis$postInit()

jmvcore::Analysis$print()

jmvcore::Analysis$readDataset()

jmvcore::Analysis$run()

jmvcore::Analysis$serialize()

jmvcore::Analysis$setError()

jmvcore::Analysis$setStatus()

ClinicoPath::agreementBase$initialize()

Method clone()

The objects of this class are cloneable with this method.

Usage

agreementClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.