Benford's Law Analysis

This function performs a Benford's Law analysis on a numeric variable. It returns the Benford's Law distribution and a list of potential suspects.

Value

A list with the Benford's Law distribution and a list of potential suspects.

A list with the Benford's Law distribution and a list of potential suspects.

Details

The Benford's Law analysis is a test to determine if the distribution of the first digits of a numeric variable follows Benford's Law. The Benford's Law distribution is compared to the observed distribution of the first digits of the variable. The analysis returns a list of potential suspects that deviate significantly from Benford's Law.

Super classes

jmvcore::Analysis -> ClinicoPath::benfordBase -> benfordClass

Methods

Public methods

benfordClass$clone()

Inherited methods

jmvcore::Analysis$.createImage()

jmvcore::Analysis$.createImages()

jmvcore::Analysis$.createPlotObject()

jmvcore::Analysis$.load()

jmvcore::Analysis$.render()

jmvcore::Analysis$.save()

jmvcore::Analysis$.savePart()

jmvcore::Analysis$.setCheckpoint()

jmvcore::Analysis$.setParent()

jmvcore::Analysis$.setReadDatasetHeaderSource()

jmvcore::Analysis$.setReadDatasetSource()

jmvcore::Analysis$.setResourcesPathSource()

jmvcore::Analysis$.setStatePathSource()

jmvcore::Analysis$addAddon()

jmvcore::Analysis$asProtoBuf()

jmvcore::Analysis$asSource()

jmvcore::Analysis$check()

jmvcore::Analysis$init()

jmvcore::Analysis$optionsChangedHandler()

jmvcore::Analysis$postInit()

jmvcore::Analysis$print()

jmvcore::Analysis$readDataset()

jmvcore::Analysis$run()

jmvcore::Analysis$serialize()

jmvcore::Analysis$setError()

jmvcore::Analysis$setStatus()

jmvcore::Analysis$translate()

ClinicoPath::benfordBase$initialize()

Method clone()

The objects of this class are cloneable with this method.

Usage

benfordClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.