Smoothly Clipped Absolute Deviation Cox Regression

Smoothly Clipped Absolute Deviation (SCAD) Cox regression for high-dimensional survival data. SCAD provides automatic variable selection with oracle properties, avoiding over-penalization of large coefficients while maintaining sparsity for small coefficients. Particularly useful for genomics and high-dimensional clinical data where interpretable variable selection is crucial.

Usage

ncvregcox(
  data,
  time,
  event,
  covariates,
  penalty = "SCAD",
  cv_folds = 10,
  lambda_type = "min",
  alpha = 1,
  gamma = 3.7,
  standardize = TRUE,
  plot_path = TRUE,
  plot_cv = TRUE,
  variable_importance = TRUE
)

Arguments

data

the data as a data frame

time

survival time variable

event

event indicator (1=event, 0=censored)

covariates

predictor variables for high-dimensional analysis

penalty

Type of penalty function for variable selection

cv_folds

Number of folds for cross-validation

lambda_type

Lambda selection criterion

alpha

Mixing parameter for elastic net (1=lasso, 0=ridge)

gamma

SCAD gamma parameter or MCP gamma parameter

standardize

Standardize covariates before fitting

plot_path

Display coefficient paths plot

plot_cv

Display cross-validation error plot

variable_importance

Calculate and display variable importance metrics

Value

A results object containing:

results$instructions					a html

results$model_summary					a table

results$selected_variables					a table

results$variable_importance					a table

results$cross_validation_results					a table

results$model_comparison					a table

results$convergence_info					a table

results$regularization_path					an image

results$cv_error_plot					an image

results$variable_selection_plot					an image

results$coefficient_comparison					an image

results$model_interpretation					a html

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$model_summary$asDF

as.data.frame(results$model_summary)

Examples

ncvregcox(
    data = data,
    time = "time",
    event = "event",
    covariates = c("x1", "x2", "x3"),
    penalty = "SCAD",
    cv_folds = 10,
    lambda_type = "min"
)
#> Error in initialize(...): unused arguments (options = <environment>, data = function (..., list = character(), package = NULL, lib.loc = NULL, verbose = getOption("verbose"), envir = .GlobalEnv, overwrite = TRUE) 
#> {
#>     fileExt <- function(x) {
#>         db <- grepl("\\.[^.]+\\.(gz|bz2|xz)$", x)
#>         ans <- sub(".*\\.", "", x)
#>         ans[db] <- sub(".*\\.([^.]+\\.)(gz|bz2|xz)$", "\\1\\2", x[db])
#>         ans
#>     }
#>     my_read_table <- function(...) {
#>         lcc <- Sys.getlocale("LC_COLLATE")
#>         on.exit(Sys.setlocale("LC_COLLATE", lcc))
#>         Sys.setlocale("LC_COLLATE", "C")
#>         read.table(...)
#>     }
#>     stopifnot(is.character(list))
#>     names <- c(as.character(substitute(list(...))[-1]), list)
#>     if (!is.null(package)) {
#>         if (!is.character(package)) stop("'package' must be a character vector or NULL")
#>     }
#>     paths <- find.package(package, lib.loc, verbose = verbose)
#>     if (is.null(lib.loc)) paths <- c(path.package(package, TRUE), if (!length(package)) getwd(), paths)
#>     paths <- unique(normalizePath(paths[file.exists(paths)]))
#>     paths <- paths[dir.exists(file.path(paths, "data"))]
#>     dataExts <- tools:::.make_file_exts("data")
#>     if (length(names) == 0) {
#>         db <- matrix(character(), nrow = 0, ncol = 4)
#>         for (path in paths) {
#>             entries <- NULL
#>             packageName <- if (file_test("-f", file.path(path, "DESCRIPTION"))) basename(path) else "."
#>             if (file_test("-f", INDEX <- file.path(path, "Meta", "data.rds"))) {
#>                 entries <- readRDS(INDEX)
#>             } else {
#>                 dataDir <- file.path(path, "data")
#>                 entries <- tools::list_files_with_type(dataDir, "data")
#>                 if (length(entries)) {
#>                   entries <- unique(tools::file_path_sans_ext(basename(entries)))
#>                   entries <- cbind(entries, "")
#>                 }
#>             }
#>             if (NROW(entries)) {
#>                 if (is.matrix(entries) && ncol(entries) == 2) db <- rbind(db, cbind(packageName, dirname(path), entries)) else warning(gettextf("data index for package %s is invalid and will be ignored", sQuote(packageName)), domain = NA, call. = FALSE)
#>             }
#>         }
#>         colnames(db) <- c("Package", "LibPath", "Item", "Title")
#>         footer <- if (missing(package)) paste0("Use ", sQuote(paste("data(package =", ".packages(all.available = TRUE))")), "\n", "to list the data sets in all *available* packages.") else NULL
#>         y <- list(title = "Data sets", header = NULL, results = db, footer = footer)
#>         class(y) <- "packageIQR"
#>         return(y)
#>     }
#>     paths <- file.path(paths, "data")
#>     for (name in names) {
#>         found <- FALSE
#>         for (p in paths) {
#>             tmp_env <- if (overwrite) envir else new.env()
#>             if (file_test("-f", file.path(p, "Rdata.rds"))) {
#>                 rds <- readRDS(file.path(p, "Rdata.rds"))
#>                 if (name %in% names(rds)) {
#>                   found <- TRUE
#>                   if (verbose) message(sprintf("name=%s:\t found in Rdata.rds", name), domain = NA)
#>                   objs <- rds[[name]]
#>                   lazyLoad(file.path(p, "Rdata"), envir = tmp_env, filter = function(x) x %in% objs)
#>                   break
#>                 } else if (verbose) message(sprintf("name=%s:\t NOT found in names() of Rdata.rds, i.e.,\n\t%s\n", name, paste(names(rds), collapse = ",")), domain = NA)
#>             }
#>             files <- list.files(p, full.names = TRUE)
#>             files <- files[grep(name, files, fixed = TRUE)]
#>             if (length(files) > 1) {
#>                 o <- match(fileExt(files), dataExts, nomatch = 100)
#>                 paths0 <- dirname(files)
#>                 paths0 <- factor(paths0, levels = unique(paths0))
#>                 files <- files[order(paths0, o)]
#>             }
#>             if (length(files)) {
#>                 for (file in files) {
#>                   if (verbose) message("name=", name, ":\t file= ...", .Platform$file.sep, basename(file), "::\t", appendLF = FALSE, domain = NA)
#>                   ext <- fileExt(file)
#>                   if (basename(file) != paste0(name, ".", ext)) found <- FALSE else {
#>                     found <- TRUE
#>                     switch(ext, R = , r = {
#>                       library("utils")
#>                       sys.source(file, chdir = TRUE, envir = tmp_env)
#>                     }, RData = , rdata = , rda = load(file, envir = tmp_env), TXT = , txt = , tab = , tab.gz = , tab.bz2 = , tab.xz = , txt.gz = , txt.bz2 = , txt.xz = assign(name, my_read_table(file, header = TRUE, as.is = FALSE), envir = tmp_env), CSV = , csv = , csv.gz = , csv.bz2 = , csv.xz = assign(name, my_read_table(file, header = TRUE, sep = ";", as.is = FALSE), envir = tmp_env), found <- FALSE)
#>                   }
#>                   if (found) break
#>                 }
#>                 if (verbose) message(if (!found) "*NOT* ", "found", domain = NA)
#>             }
#>             if (found) break
#>         }
#>         if (!found) {
#>             warning(gettextf("data set %s not found", sQuote(name)), domain = NA)
#>         } else if (!overwrite) {
#>             for (o in ls(envir = tmp_env, all.names = TRUE)) {
#>                 if (exists(o, envir = envir, inherits = FALSE)) warning(gettextf("an object named %s already exists and will not be overwritten", sQuote(o))) else assign(o, get(o, envir = tmp_env, inherits = FALSE), envir = envir)
#>             }
#>             rm(tmp_env)
#>         }
#>     }
#>     invisible(names)
#> })