Interrater Reliability

Function for Interrater Reliability.

Usage

agreement(
  data,
  vars,
  sft = FALSE,
  heatmap = FALSE,
  heatmapDetails = FALSE,
  wght = "unweighted",
  exct = FALSE,
  kripp = FALSE,
  krippMethod = "nominal",
  bootstrap = FALSE
)

Arguments

data

The data as a data frame. The data should be in long format, where each row is a unique observation.

vars

A string naming the variable from data that contains the diagnosis given by the observer, variable can be categorical or ordinal.

sft

Boolean selection whether to show frequency table. Default is 'false'. If 'true', the function will show frequency table for each observer.

heatmap

Boolean selection whether to show agreement heatmap. Default is 'true'.

heatmapDetails

Boolean selection whether to show detailed heatmap with kappa values. Default is 'false'.

wght

A list for the argument weight (wght), for weighted kappa analysis. Default is 'unweighted'. 'squared' or 'equal' should be selected only with ordinal variables. The function gives error if the variable type is not ordinal. The default is 'unweighted'.

exct

Boolean selection whether to use exact kappa. Effects only more than 3 observers. Default is 'false'.

kripp

Boolean selection whether to calculate Krippendorff's alpha. Default is 'false'.

krippMethod

Specifies the measurement level for Krippendorff's alpha calculation.

bootstrap

Calculate bootstrap confidence intervals for Krippendorff's alpha.

Value

A results object containing:

results$irrtable					a table

results$heatmapPlot					an image

results$text2					a html

results$text					a preformatted

results$krippTable					a table

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$irrtable$asDF

as.data.frame(results$irrtable)

Examples

# \donttest{
# example will be added
# }