Interrater Reliability

Function for Interrater Reliability.

agreement(data, vars, sft = FALSE, wght = "unweighted", exct = FALSE)

Arguments

data
The data as a data frame.

vars
A string naming the variable from data that contains the diagnosis given by the observer, variable can be categorical, ordinal or numeric.

sft
Boolean selection whether to show frequency table. Default is 'false'.

wght
A list for the argument weight (wght), for weighted kappa analysis. Default is 'unweighted'. 'squared' or 'equal' should be selected only with ordinal variables. The function gives error if the variable type is not ordinal.

exct
Boolean selection whether to use exact kappa. Effects only more than 3 observers.

Value

A results object containing:

results$text					a preformatted

results$irrtable					a table

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$irrtable$asDF

as.data.frame(results$irrtable)

Examples

if (FALSE) {
# example will be added
}