Parsnip Survival Model Wrappers

Unified interface to survival models using the parsnip package and tidymodels framework. Supports multiple survival model engines with standardized syntax for model fitting, prediction, and evaluation.

Usage

jparsnipsurvival(
  data,
  time_var,
  event_var,
  predictors,
  model_engine = "cox_survival",
  parametric_dist = "weibull",
  prediction_type = "linear_pred",
  penalty = 0.01,
  mixture = 1,
  trees = 500,
  min_n = 10,
  split_data = FALSE,
  train_prop = 0.8,
  cross_validation = FALSE,
  cv_folds = 5,
  bootstrap_resamples = 100,
  show_predictions = TRUE,
  show_metrics = TRUE,
  show_coefficients = TRUE,
  show_residuals = FALSE,
  confidence_level = 0.95
)

Arguments

data

Dataset for survival modeling

time_var

Survival time or time to event variable

event_var

Event indicator (0=censored, 1=event)

predictors

Variables to include in the survival model

model_engine

Survival model engine to use

parametric_dist

Distribution for parametric survival models

prediction_type

Type of prediction to generate

penalty

L1/L2 penalty parameter for regularized models

mixture

Elastic net mixing parameter (0=ridge, 1=lasso)

trees

Number of trees for random forest models

min_n

Minimum number of observations in terminal nodes

split_data

Split data into training and testing sets

train_prop

Proportion of data to use for training

cross_validation

Perform cross-validation for model assessment

cv_folds

Number of cross-validation folds

bootstrap_resamples

Number of bootstrap resamples for uncertainty estimation

show_predictions

Display model predictions

show_metrics

Display model performance metrics

show_coefficients

Display model coefficients and statistics

show_residuals

Display residual plots and diagnostics

confidence_level

Confidence level for intervals

Value

A results object containing:

results$instructions					a html

results$model_info					a table

results$coefficients					a table

results$predictions					a table

results$performance_metrics					a table

results$cross_validation_results					a table

results$variable_importance					a table

results$survival_curves					an image

results$residual_plots					an image

results$model_summary					a html

results$interpretation					a html

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$model_info$asDF

as.data.frame(results$model_info)